requirements-old.txt
alembic==1.0.10
aniso8601==7.0.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
kitchen==1.2.6
lockfile==0.12.2
Mako==1.0.12
MarkupSafe==1.1.1
munch==2.3.2
openidc-client==0.6.0
psycopg2==2.8.3
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
python-fedora==0.10.0
python-openid-cla==1.2
python-openid-teams==1.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.4
urllib3==1.25.3
Werkzeug==0.15.4
